requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.12.0
shap>=0.45.0
torch>=2.2.0
torchvision>=0.17.0
sentence-transformers>=2.6.0
transformers>=4.39.0
statsmodels>=0.14.0
optuna>=3.6.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
scipy>=1.13.0
joblib>=1.3.0
pillow>=10.2.0
redis>=5.0.0
sqlalchemy>=2.0.0
pytest>=8.1.0
httpx>=0.27.0
python-multipart>=0.0.9
